+++ CHANGES.rst
@@ -1,6 +1,10 @@
 1.1 (unreleased)
 ----------------
 
+* Support surrogate pairs by using math text-run groups.
+* Drop support for Python 3.1 and 3.2; 3.3 - 3.5 are now officially supported.
+
+
 1.0 (2012-06-06)
 ----------------